requirements.txt
django~=5.1.4
django-colorfield~=0.11.0
django-guardian~=2.4.0
djangorestframework~=3.15.2
djangorestframework-guardian2~=0.7.0
drf-nested-routers~=0.94.1
drf_redesign~=0.4.0
django-filter~=24.3
markdown~=3.7